export lean index.ts interface
