+++ vscode/microsoft-kiota/src/kiotaInterop/lib/generateClient.ts
@@ -2,7 +2,7 @@ import * as rpc from "vscode-jsonrpc/node";
 
 import { ConsumerOperation, GenerationConfiguration, KiotaLogEntry } from "..";
 import connectToKiota from "../connect";
-import { KiotaGenerationLanguage } from "../enums";
+import { KiotaGenerationLanguage } from "../types";
 
 interface ClientGenerationOptions {
   clearCache: boolean;
