+++ vscode/microsoft-kiota/src/providers/dependenciesViewProvider.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
 
-import { DependencyType, dependencyTypeToString, generationLanguageToString, LanguageInformation, LanguagesInformation } from '../kiotaInterop';
-import { KiotaGenerationLanguage } from '../kiotaInterop/enums';
+import { DependencyType, dependencyTypeToString, generationLanguageToString, KiotaGenerationLanguage, LanguageInformation, LanguagesInformation } from '../kiotaInterop';
 
 export class DependenciesViewProvider implements vscode.WebviewViewProvider {
     private _view?: vscode.WebviewView;
